requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==10.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pdf2image==1.16.3